requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
graphql-core==3.2.3
h11==0.14.0
idna==3.6
libcst==1.3.1
markdown-it-py==3.0.0
mdurl==0.1.2
pip==24.0
pydantic_core==2.16.3
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
strawberry-graphql==0.224.1
typer==0.12.1
typing_extensions==4.11.0
urllib3==2.2.1
fastapi==0.110.1
pydantic==2.6.4
uvicorn==0.29.0